Add dispatch to popups - error in add service and popup from multiple links

+++ backend/src/main/java/com/example/madrid/madrid/Restaurant.java
@@ -16,8 +16,8 @@ public class Restaurant {
     private String item;
     private Double price;
     private String description;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
     // Getters and setters
     
@@ -61,19 +61,19 @@ public class Restaurant {
         this.description = description;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
 }
